Introduce a few more options for tuning log output.
Remove connections from tracking_conns on connection_state_remove, and
require that the connection gets one final log line. In order to address
the issue that that final line could be a possible duplicate, and to
help users tune noise in long-idle connections, introduce an option to
not log updates when no bytes have been transferred since the last line.